%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  File.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

 \name{as.character.File}
\alias{as.character.File}
\alias{File.as.character}
\alias{as.character.File}
\alias{as.character,File-method}


 \title{Gets a character string representation of the File object}

 \usage{\method{as.character}{File}(x, ...)}

 \description{
  Gets a character string representation of the File object, which is the absolute path as returned by
  \code{getAbsolutePath()}.
 }

 \value{
   Returns a \code{\link[base]{character}} string.
 }



 \author{Henrik Bengtsson (\url{http://www.braju.com/R/})}

 \seealso{
   \code{\link[R.io:getAbsolutePath.File]{*getAbsolutePath}()}.
   For more information see \code{\link{File}}.
 }
\keyword{internal}
\keyword{methods}
