%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  RegExprFileFilter.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{as.character.RegExprFileFilter}
\alias{as.character.RegExprFileFilter}
\alias{RegExprFileFilter.as.character}
\alias{as.character.RegExprFileFilter}
\alias{as.character,RegExprFileFilter-method}


\title{Gets a string representation of the regular expression file filter}

\usage{\method{as.character}{RegExprFileFilter}(x, ...)}

\description{
 Gets a string representation of the regular expression file filter.
}

\value{
  Returns a \code{\link[base]{character}} string.
}

\author{Henrik Bengtsson (\url{http://www.braju.com/R/})}

\seealso{
  For more information see \code{\link{RegExprFileFilter}}..
}
\keyword{internal}
\keyword{methods}
