%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  FileInputStream.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

 \name{available.FileInputStream}
\alias{available.FileInputStream}
\alias{FileInputStream.available}
\alias{available.FileInputStream}
\alias{available,FileInputStream-method}


 \title{Returns the number of bytes currently available in input buffer}

 \usage{\method{available}{FileInputStream}(this, ...)}

 \description{
  Returns the number of bytes currently available in input buffer, which is the minimum number of bytes that can be read
   without blocking (having to wait).
 }

 \value{
   Returns an \code{\link[base]{integer}} greater or equal to zero.
 }



 \author{Henrik Bengtsson (\url{http://www.braju.com/R/})}

 \seealso{
  For more information see \code{\link{FileInputStream}}.
 }
\keyword{internal}
\keyword{methods}
