%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  ConnectionInputStream.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{close.ConnectionInputStream}
\alias{close.ConnectionInputStream}
\alias{ConnectionInputStream.close}
\alias{close.ConnectionInputStream}
\alias{close,ConnectionInputStream-method}


\title{Closes the input stream and the connection}

\usage{\method{close}{ConnectionInputStream}(con, ...)}

\description{
  Closes the input stream and the connection.
}

\value{
  Returns nothing.
}

\examples{\dontrun{See example in help(ConnectionInputStream) for an example.}}

\author{Henrik Bengtsson (\url{http://www.braju.com/R/})}

\seealso{
  For more information see \code{\link{ConnectionInputStream}}.
}
\keyword{internal}
\keyword{methods}
