%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  FileInputStream.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{close.FileInputStream}
\alias{close.FileInputStream}
\alias{FileInputStream.close}
\alias{close.FileInputStream}
\alias{close,FileInputStream-method}


\title{Closes the input stream and releases the file}

\usage{\method{close}{FileInputStream}(con, ...)}

\description{
 Closes the input stream and releases the file.
}

\value{
  Returns nothing.
}

\author{Henrik Bengtsson (\url{http://www.braju.com/R/})}

\seealso{
 For more information see \code{\link{FileInputStream}}.
}
\keyword{internal}
\keyword{methods}
