%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  FileOutputStream.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

 \name{close.FileOutputStream}
\alias{close.FileOutputStream}
\alias{FileOutputStream.close}
\alias{close.FileOutputStream}
\alias{close,FileOutputStream-method}


 \title{Flushes and closes the file output stream}

 \usage{\method{close}{FileOutputStream}(con, ...)}

 \description{
  Flushes and closes the file output stream.
 }

 \value{
   Returns nothing.
 }



 \author{Henrik Bengtsson (\url{http://www.braju.com/R/})}

 \seealso{
  For more information see \code{\link{FileOutputStream}}.
 }
\keyword{internal}
\keyword{methods}
