%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  InputStream.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{close.InputStream}
\alias{close.InputStream}
\alias{InputStream.close}
\alias{close.InputStream}
\alias{close,InputStream-method}


\title{Closes the input stream}

\usage{\method{close}{InputStream}(con, ...)}

\description{
 Closes the input stream.
}

\value{
  Returns nothing.
}

\author{Henrik Bengtsson (\url{http://www.braju.com/R/})}

\seealso{
 For more information see \code{\link{InputStream}}.
}
\keyword{internal}
\keyword{methods}
