%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  MultiOutputStream.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{close.MultiOutputStream}
\alias{close.MultiOutputStream}
\alias{MultiOutputStream.close}
\alias{close.MultiOutputStream}
\alias{close,MultiOutputStream-method}


\title{Closes all output streams}

\usage{\method{close}{MultiOutputStream}(con, ...)}

\description{
 Closes all output streams.
}

\value{
  Returns nothing.
}

\seealso{
  For more information see \code{\link{MultiOutputStream}}.
}

\author{Henrik Bengtsson (\url{http://www.braju.com/R/})}
\keyword{internal}
\keyword{methods}
