%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  OutputStream.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{close.OutputStream}
\alias{close.OutputStream}
\alias{OutputStream.close}
\alias{close.OutputStream}
\alias{close,OutputStream-method}


\title{Closes the output stream}

\usage{\method{close}{OutputStream}(con, ...)}

\description{
 Closes the output stream, but first the output stream is flushed.
}

\value{
  Returns nothing.
}

\seealso{
  For more information see \code{\link{OutputStream}}.
}

\author{Henrik Bengtsson (\url{http://www.braju.com/R/})}
\keyword{internal}
\keyword{methods}
