%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  FileOutputStream.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

 \name{finalize.FileOutputStream}
\alias{finalize.FileOutputStream}
\alias{FileOutputStream.finalize}
\alias{finalize.FileOutputStream}
\alias{finalize,FileOutputStream-method}


 \title{Finalizes the stream by first closing it}

 \usage{\method{finalize}{FileOutputStream}(this, ...)}

 \description{
  Finalizes the stream by first closing it.
 }

 \value{
   Returns nothing.
 }



 \author{Henrik Bengtsson (\url{http://www.braju.com/R/})}

 \seealso{
  For more information see \code{\link{FileOutputStream}}.
 }
\keyword{internal}
\keyword{methods}
