%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  InputStream.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{finalize.InputStream}
\alias{finalize.InputStream}
\alias{InputStream.finalize}
\alias{finalize.InputStream}
\alias{finalize,InputStream-method}


\title{Finalizes the stream by closing it}

\usage{\method{finalize}{InputStream}(this, ...)}

\description{
 Finalizes the stream by closing it.
}

\author{Henrik Bengtsson (\url{http://www.braju.com/R/})}

\seealso{
  For more information see \code{\link{InputStream}}.
}
\keyword{internal}
\keyword{methods}
