%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  OutputStream.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{finalize.OutputStream}
\alias{finalize.OutputStream}
\alias{OutputStream.finalize}
\alias{finalize.OutputStream}
\alias{finalize,OutputStream-method}


\title{Finalize method that closes the stream when it is deleted}

\usage{\method{finalize}{OutputStream}(this, ...)}

\description{
 Finalize method that closes the stream when it is deleted.
}

\seealso{
  For more information see \code{\link{OutputStream}}.
}

\author{Henrik Bengtsson (\url{http://www.braju.com/R/})}
\keyword{internal}
\keyword{methods}
