%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  Reporter.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{finalize.Reporter}
\alias{finalize.Reporter}
\alias{Reporter.finalize}
\alias{finalize.Reporter}
\alias{finalize,Reporter-method}


\title{Finalizes the reporter by first closing it}

\usage{\method{finalize}{Reporter}(this, ...)}

\description{
  Finalizes the reporter by first closing it.
}

\author{Henrik Bengtsson (\url{http://www.braju.com/R/})}

\seealso{
  For more information see \code{\link{Reporter}}.
}
\keyword{internal}
\keyword{methods}
