%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  OutputStream.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{flush.OutputStream}
\alias{flush.OutputStream}
\alias{OutputStream.flush}
\alias{flush.OutputStream}
\alias{flush,OutputStream-method}


\title{Flushes the output stream}

\usage{\method{flush}{OutputStream}(con, ...)}

\description{
 Flushes the output stream and emptying any internal buffers.
}

\value{
  Returns nothing.
}

\seealso{
  For more information see \code{\link{OutputStream}}.
}

\author{Henrik Bengtsson (\url{http://www.braju.com/R/})}
\keyword{internal}
\keyword{methods}
