%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  File.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

 \name{getAbsoluteFile.File}
\alias{getAbsoluteFile.File}
\alias{File.getAbsoluteFile}
\alias{getAbsoluteFile.File}
\alias{getAbsoluteFile,File-method}


 \title{Gets the absolute path as a File}

 \usage{\method{getAbsoluteFile}{File}(this, ...)}

 \description{
   Gets the absolute path as a File.
 }

 \arguments{
  \item{...}{Additional arguments accepted by \code{\link[R.io:getAbsolutePath.File]{*getAbsolutePath}()}.}
 }

 \value{
  Returns a \code{File} object of the absolute pathname.
 }



 \author{Henrik Bengtsson (\url{http://www.braju.com/R/})}

 \seealso{
   For more information see \code{\link{File}}.
 }
\keyword{internal}
\keyword{methods}
