%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  File.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

 \name{getCanonicalFile.File}
\alias{getCanonicalFile.File}
\alias{File.getCanonicalFile}
\alias{getCanonicalFile.File}
\alias{getCanonicalFile,File-method}


 \title{Gets the cannonical form of this pathname}

 \usage{\method{getCanonicalFile}{File}(this, ...)}

 \description{
   Gets the cannonical form of this pathname.
 }

 \value{
  Returns a \code{File} object.
 }



 \author{Henrik Bengtsson (\url{http://www.braju.com/R/})}

 \seealso{
   For more information see \code{\link{File}}.
 }
\keyword{internal}
\keyword{methods}
