%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  File.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

 \name{getCanonicalPath.File}
\alias{getCanonicalPath.File}
\alias{File.getCanonicalPath}
\alias{getCanonicalPath.File}
\alias{getCanonicalPath,File-method}


 \title{Gets the cannonical pathname string}

 \usage{\method{getCanonicalPath}{File}(this, ...)}

 \description{
   Gets the cannonical pathname string.
 }

 \value{
  Returns a \code{\link[base]{character}} string.
 }



 \author{Henrik Bengtsson (\url{http://www.braju.com/R/})}

 \seealso{
   For more information see \code{\link{File}}.
 }
\keyword{internal}
\keyword{methods}
