%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  FileException.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{getFile.FileException}
\alias{getFile.FileException}
\alias{FileException.getFile}
\alias{getFile.FileException}
\alias{getFile,FileException-method}


\title{Gets the file that failed to be opened}

\usage{\method{getFile}{FileException}(this, ...)}

\description{
 Gets the file that failed to be opened.
}

\value{
  Returns a \code{\link[base]{character}} string.
}

\author{Henrik Bengtsson (\url{http://www.braju.com/R/})}

\seealso{
  To get the pathname as a \code{\link[base]{character}} string, see \code{\link[R.io:getPathname.FileException]{*getPathname}()}.
  For more information see \code{\link{FileException}}.
}
\keyword{internal}
\keyword{methods}
