%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  RegExprFileFilter.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{getMask.RegExprFileFilter}
\alias{getMask.RegExprFileFilter}
\alias{RegExprFileFilter.getMask}
\alias{getMask.RegExprFileFilter}
\alias{getMask,RegExprFileFilter-method}


\title{Gets the regular expression mask currently used}

\usage{\method{getMask}{RegExprFileFilter}(this, ...)}

\description{
 Gets the regular expression mask currently used.
}

\value{
  Returns a \code{\link[base]{character}} string.
}

\author{Henrik Bengtsson (\url{http://www.braju.com/R/})}

\seealso{
  \code{\link[R.io:getMask.RegExprFileFilter]{*getMask}()}.
  For more information see \code{\link{RegExprFileFilter}}..
}
\keyword{internal}
\keyword{methods}
