%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  File.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

 \name{getName.File}
\alias{getName.File}
\alias{File.getName}
\alias{getName.File}
\alias{getName,File-method}


 \title{Gets the name or directory specified by this pathname}

 \usage{\method{getName}{File}(this, removeSuffix=FALSE, ...)}

 \description{
  Gets the name or directory specified by this pathname.
  This is basically, the non-empty sequence of characters after the last
  path separator.
 }

 \value{
  Returns a \code{\link[base]{character}} string.
 }



 \author{Henrik Bengtsson (\url{http://www.braju.com/R/})}

 \seealso{
   \code{\link[base]{basename}}().
   For more information see \code{\link{File}}.
 }
\keyword{internal}
\keyword{methods}
