%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  File.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

 \name{getParentFile.File}
\alias{getParentFile.File}
\alias{File.getParentFile}
\alias{getParentFile.File}
\alias{getParentFile,File-method}


 \title{Gets the parent specified by this pathname}

 \usage{\method{getParentFile}{File}(this, ...)}

 \description{
  Gets the parent specified by this pathname.
 }

 \value{
  Returns a \code{File} object if the parent exists, otherwise \code{\link[base]{NULL}}.
 }



 \author{Henrik Bengtsson (\url{http://www.braju.com/R/})}

 \seealso{
   \code{\link[R.io:getParent.File]{*getParent}()}
   For more information see \code{\link{File}}.
 }
\keyword{internal}
\keyword{methods}
