%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  File.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

 \name{getPath.File}
\alias{getPath.File}
\alias{File.getPath}
\alias{getPath.File}
\alias{getPath,File-method}


 \title{Gets the path specified by this pathname}

 \usage{\method{getPath}{File}(this, removeSuffix=FALSE, ...)}

 \description{
  Gets the path specified by this pathname using system specific file
  separators.
 }

 \value{
  Returns a \code{\link[base]{character}} string.
 }



 \author{Henrik Bengtsson (\url{http://www.braju.com/R/})}

 \seealso{
   For more information see \code{\link{File}}.
 }
\keyword{internal}
\keyword{methods}
