%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  FileException.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{getPathname.FileException}
\alias{getPathname.FileException}
\alias{FileException.getPathname}
\alias{getPathname.FileException}
\alias{getPathname,FileException-method}


\title{Gets the pathname of the file that failed to be opened}

\usage{\method{getPathname}{FileException}(this, ...)}

\description{
 Gets the pathname of the file that failed to be opened.
}

\value{
  Returns a \code{\link[base]{character}} string.
}

\author{Henrik Bengtsson (\url{http://www.braju.com/R/})}

\seealso{
  To get the pathname as a \code{\link{File}} object, see \code{\link[R.io:getFile.FileException]{*getFile}()}.
  For more information see \code{\link{FileException}}.
}
\keyword{internal}
\keyword{methods}
