%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  FileNotFoundException.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{getPattern.FileNotFoundException}
\alias{getPattern.FileNotFoundException}
\alias{FileNotFoundException.getPattern}
\alias{getPattern.FileNotFoundException}
\alias{getPattern,FileNotFoundException-method}


\title{Gets the filename pattern that was not matched}

\usage{\method{getPattern}{FileNotFoundException}(this, ...)}

\description{
 Gets the filename pattern that was not matched.
}

\value{
  Returns a \code{\link[base]{character}} string.
}

\author{Henrik Bengtsson (\url{http://www.braju.com/R/})}

\seealso{
  For more information see \code{\link{FileNotFoundException}}.
}
\keyword{internal}
\keyword{methods}
