%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  File.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

 \name{getWithoutExtension.File}
\alias{getWithoutExtension.File}
\alias{File.getWithoutExtension}
\alias{getWithoutExtension.File}
\alias{getWithoutExtension,File-method}


 \title{Takes a filename and returns the filename without the extension}

 \synopsis{getWithoutExtension.File(this, filename, ...)}
\usage{File$getWithoutExtension(filename, ...)}

 \description{
   Takes a filename and returns the filename without the extension.
 }

 \value{Returns a \code{\link[base]{character}} string.}

 \author{Henrik Bengtsson (\url{http://www.braju.com/R/})}

 \seealso{
   For more information see \code{\link{File}}.
 }
\keyword{internal}
\keyword{methods}
