%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  File.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

 \name{isAbsolute.File}
\alias{isAbsolute.File}
\alias{File.isAbsolute}
\alias{isAbsolute.File}
\alias{isAbsolute,File-method}


 \title{Checks if this pathname is absolute}

 \usage{\method{isAbsolute}{File}(this, ...)}

 \description{
   Checks if this pathname is absolute.
 }

 \value{
  Returns a \code{\link[base:logical]{TRUE}} if the pathname is absolute, otherwise \code{\link[base:logical]{FALSE}}.
 }



 \author{Henrik Bengtsson (\url{http://www.braju.com/R/})}

 \seealso{
   For more information see \code{\link{File}}.
 }
\keyword{internal}
\keyword{methods}
