%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  Reporter.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{isCreator.Reporter}
\alias{isCreator.Reporter}
\alias{Reporter.isCreator}
\alias{isCreator.Reporter}
\alias{isCreator,Reporter-method}


\title{Checks if reporter is a creator or not}

\usage{\method{isCreator}{Reporter}(this, ...)}

\description{
  Checks if reporter is a creator or not.
}

\value{
  Returns \code{\link[base:logical]{TRUE}} if the reporter is a creator, otherwise \code{\link[base:logical]{FALSE}}.
}

\author{Henrik Bengtsson (\url{http://www.braju.com/R/})}

\seealso{
  For more information see \code{\link{Reporter}}.
}
\keyword{internal}
\keyword{methods}
