%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  InputStream.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{mark.InputStream}
\alias{mark.InputStream}
\alias{InputStream.mark}
\alias{mark.InputStream}
\alias{mark,InputStream-method}


\title{Marks the current position}

\usage{\method{mark}{InputStream}(this, ...)}

\description{
 Marks the current position which can be returned to by
 \code{reset()}.
}

\value{
  Returns nothing.
}

\author{Henrik Bengtsson (\url{http://www.braju.com/R/})}

\seealso{
 \code{\link[R.io:markSupported.InputStream]{*markSupported}()} and \code{\link[R.io:reset.InputStream]{*reset}()}.
 For more information see \code{\link{InputStream}}.
}
\keyword{internal}
\keyword{methods}
