%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  InputStream.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{markSupported.InputStream}
\alias{markSupported.InputStream}
\alias{InputStream.markSupported}
\alias{markSupported.InputStream}
\alias{markSupported,InputStream-method}


\title{Checks if the input stream supports mark() and reset()}

\usage{\method{markSupported}{InputStream}(this, ...)}

\description{
 Checks if the input stream supports \code{mark()} and \code{reset()}.
}

\value{
  Returns a logical.
}

\author{Henrik Bengtsson (\url{http://www.braju.com/R/})}

\seealso{
 \code{\link[R.io:mark.InputStream]{*mark}()} and \code{\link[R.io:reset.InputStream]{*reset}()}.
 For more information see \code{\link{InputStream}}.
}
\keyword{internal}
\keyword{methods}
