%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  999.NonDocumentedObjects.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{Non-documented objects}
\alias{Non-documented objects}
\title{Non-documented objects}


\alias{accept}
\alias{add}
\alias{add.MultiReporter}
\alias{as.character.ConnectionInputStream}
\alias{as.character.ConnectionOutputStream}
\alias{as.character.RspException}
\alias{as.HTML}
\alias{as.HTML.RspException}
\alias{asByte}
\alias{asByte.Java}
\alias{asInt}
\alias{asInt.Java}
\alias{asLong}
\alias{asLong.Java}
\alias{asShort}
\alias{asShort.Java}
\alias{available}
\alias{available.BufferedOutputStream}
\alias{available.FilterInputStream}
\alias{available.LineNumberInputStream}
\alias{beginBold}
\alias{beginBold.HtmlReporter}
\alias{beginBold.LaTeXReporter}
\alias{beginBold.MultiReporter}
\alias{beginBold.Reporter}
\alias{beginComment}
\alias{beginComment.LaTeXReporter}
\alias{beginComment.Reporter}
\alias{beginComment.TextReporter}
\alias{beginDebug}
\alias{beginDebug.HtmlReporter}
\alias{beginDebug.LaTeXReporter}
\alias{beginDebug.Reporter}
\alias{beginDebug.TextReporter}
\alias{beginEmphased}
\alias{beginEmphased.HtmlReporter}
\alias{beginEmphased.LaTeXReporter}
\alias{beginEmphased.MultiReporter}
\alias{beginEmphased.Reporter}
\alias{beginEnvironment}
\alias{beginEnvironment.LaTeXReporter}
\alias{beginEquation}
\alias{beginEquation.HtmlReporter}
\alias{beginEquation.LaTeXReporter}
\alias{beginEquation.Reporter}
\alias{beginEquation.TextReporter}
\alias{beginFigure}
\alias{beginFigure.HtmlReporter}
\alias{beginFigure.LaTeXReporter}
\alias{beginFigure.MultiReporter}
\alias{beginFigure.Reporter}
\alias{beginFigure.TextReporter}
\alias{beginItalic}
\alias{beginItalic.HtmlReporter}
\alias{beginItalic.LaTeXReporter}
\alias{beginItalic.MultiReporter}
\alias{beginItalic.Reporter}
\alias{beginParagraph}
\alias{beginParagraph.HtmlReporter}
\alias{beginParagraph.LaTeXReporter}
\alias{beginParagraph.MultiReporter}
\alias{beginParagraph.Reporter}
\alias{beginTable}
\alias{beginTable.MultiReporter}
\alias{beginTag}
\alias{beginTag.LaTeXReporter}
\alias{beginVerbatim}
\alias{beginVerbatim.HtmlReporter}
\alias{beginVerbatim.LaTeXReporter}
\alias{beginVerbatim.MultiReporter}
\alias{beginVerbatim.Reporter}
\alias{canRead}
\alias{canWrite}
\alias{chooseFile}
\alias{chooseFile.File}
\alias{close.BufferedOutputStream}
\alias{close.ConnectionOutputStream}
\alias{close.FilterInputStream}
\alias{close.FilterOutputStream}
\alias{close.HtmlReporter}
\alias{close.LaTeXReporter}
\alias{close.LineNumberInputStream}
\alias{close.MultiReporter}
\alias{close.TextReporter}
\alias{ConnectionOutputStream}
\alias{contains}
\alias{contains.MultiReporter}
\alias{createNewFile}
\alias{createTempFile}
\alias{currentImageExists}
\alias{currentImageExists.HtmlRotatableImage}
\alias{definition}
\alias{endBold}
\alias{endBold.HtmlReporter}
\alias{endBold.LaTeXReporter}
\alias{endBold.MultiReporter}
\alias{endBold.Reporter}
\alias{endComment}
\alias{endComment.LaTeXReporter}
\alias{endComment.Reporter}
\alias{endComment.TextReporter}
\alias{endDebug}
\alias{endDebug.HtmlReporter}
\alias{endDebug.LaTeXReporter}
\alias{endDebug.Reporter}
\alias{endDebug.TextReporter}
\alias{endEmphased}
\alias{endEmphased.HtmlReporter}
\alias{endEmphased.LaTeXReporter}
\alias{endEmphased.MultiReporter}
\alias{endEmphased.Reporter}
\alias{endEnvironment}
\alias{endEnvironment.LaTeXReporter}
\alias{endEquation}
\alias{endEquation.HtmlReporter}
\alias{endEquation.LaTeXReporter}
\alias{endEquation.Reporter}
\alias{endEquation.TextReporter}
\alias{endFigure}
\alias{endFigure.HtmlReporter}
\alias{endFigure.LaTeXReporter}
\alias{endFigure.MultiReporter}
\alias{endFigure.Reporter}
\alias{endFigure.TextReporter}
\alias{endItalic}
\alias{endItalic.HtmlReporter}
\alias{endItalic.LaTeXReporter}
\alias{endItalic.MultiReporter}
\alias{endItalic.Reporter}
\alias{endParagraph}
\alias{endParagraph.HtmlReporter}
\alias{endParagraph.LaTeXReporter}
\alias{endParagraph.MultiReporter}
\alias{endParagraph.Reporter}
\alias{endParagraph.TextReporter}
\alias{endsWith}
\alias{endsWith.default}
\alias{endTable}
\alias{endTable.MultiReporter}
\alias{endTag}
\alias{endTag.LaTeXReporter}
\alias{endVerbatim}
\alias{endVerbatim.HtmlReporter}
\alias{endVerbatim.LaTeXReporter}
\alias{endVerbatim.MultiReporter}
\alias{endVerbatim.Reporter}
\alias{erase}
\alias{escape}
\alias{escapeText}
\alias{escapeText.LaTeXReporter}
\alias{FilterInputStream}
\alias{FilterOutputStream}
\alias{finalize.MultiReporter}
\alias{flush.BufferedOutputStream}
\alias{flush.FilterOutputStream}
\alias{getAbsoluteFile}
\alias{getAbsolutePath}
\alias{getCanonicalFile}
\alias{getCanonicalPath}
\alias{getCode}
\alias{getCode.RspException}
\alias{getCurrentConnection}
\alias{getCurrentConnection.RspEngine}
\alias{getCurrentPath}
\alias{getCurrentPath.RspEngine}
\alias{getExtension}
\alias{getExtension.HtmlReporter}
\alias{getExtension.LaTeXReporter}
\alias{getExtension.MultiReporter}
\alias{getExtension.Reporter}
\alias{getExtension.TextReporter}
\alias{getFigureFilename}
\alias{getFigureFilename.Reporter}
\alias{getFigureNameFormat}
\alias{getFigureNameFormat.MultiReporter}
\alias{getFigureNameFormat.Reporter}
\alias{getFile}
\alias{getHtmlImgTag}
\alias{getHtmlImgTag.HtmlRotatableImage}
\alias{getHtmlMapTag}
\alias{getHtmlMapTag.HtmlRotatableImage}
\alias{getHtmlPrintStream}
\alias{getHtmlPrintStream.RspEngine}
\alias{getHtmlScriptTag}
\alias{getHtmlScriptTag.HtmlRotatableImage}
\alias{getImageName}
\alias{getImageName.HtmlRotatableImage}
\alias{getImageNames}
\alias{getImageNames.HtmlRotatableImage}
\alias{getLineNumber}
\alias{getLineNumber.LineNumberInputStream}
\alias{getMapName}
\alias{getMapName.HtmlRotatableImage}
\alias{getMask}
\alias{getMessage.FileExistsException}
\alias{getMessage.FileNotFoundException}
\alias{getOnStart}
\alias{getOnStart.LaTeXReporter}
\alias{getOutput}
\alias{getOutput.RspEngine}
\alias{getOutputPath}
\alias{getOutputPath.RspEngine}
\alias{getParent}
\alias{getParentFile}
\alias{getPathname}
\alias{getPattern}
\alias{getPreloadJSCode}
\alias{getPreloadJSCode.HtmlRotatableImage}
\alias{getProgress}
\alias{getProgress.HtmlRotatableImage}
\alias{getset}
\alias{getVerbatim}
\alias{getVerbatim.Reporter}
\alias{getVerbatimOld}
\alias{getVerbatimOld.Reporter}
\alias{getWithoutExtension}
\alias{hasEndTag}
\alias{HtmlReporter}
\alias{include}
\alias{include.MultiReporter}
\alias{include.Reporter}
\alias{indent}
\alias{isAbsolute}
\alias{isCloseable}
\alias{isCloseable.ConnectionInputStream}
\alias{isCloseable.ConnectionOutputStream}
\alias{isCreator}
\alias{isDirectory}
\alias{isExisting}
\alias{isFile}
\alias{isMarkSupported}
\alias{isMarkSupported.FilterInputStream}
\alias{isOpen}
\alias{isOpen.ConnectionInputStream}
\alias{isOpen.ConnectionOutputStream}
\alias{isOpen.default}
\alias{lastModified}
\alias{LaTeXReporter}
\alias{LineNumberInputStream}
\alias{listDir}
\alias{listFiles}
\alias{main}
\alias{main.MultiReporter}
\alias{mark}
\alias{mark.FilterInputStream}
\alias{mark.LineNumberInputStream}
\alias{markSupported}
\alias{method}
\alias{mkdir}
\alias{mkdirs}
\alias{MultiReporter}
\alias{nbrOfImages}
\alias{nbrOfImages.HtmlRotatableImage}
\alias{ncol}
\alias{ncol.default}
\alias{ncol.HtmlRotatableImage}
\alias{nrow}
\alias{nrow.default}
\alias{nrow.HtmlRotatableImage}
\alias{open.HtmlReporter}
\alias{open.LaTeXReporter}
\alias{open.MultiReporter}
\alias{open.TextReporter}
\alias{parseAttributes}
\alias{parseAttributes.RspEngine}
\alias{popRedirect}
\alias{popRedirect.MultiReporter}
\alias{popRedirect.Reporter}
\alias{popTag}
\alias{popTags}
\alias{prefix}
\alias{prepareTable}
\alias{prepareTable.Reporter}
\alias{print.LaTeXReporter}
\alias{printAsIs}
\alias{printAsIs.LaTeXReporter}
\alias{printComment}
\alias{printComment.LaTeXReporter}
\alias{printDl}
\alias{printEnvironment}
\alias{printEnvironment.LaTeXReporter}
\alias{printList}
\alias{println}
\alias{println.LaTeXReporter}
\alias{printOl}
\alias{printTable}
\alias{printTag}
\alias{printTag.LaTeXReporter}
\alias{printTagLn}
\alias{printTagLn.LaTeXReporter}
\alias{printTimestamp}
\alias{printUl}
\alias{process}
\alias{process.RspEngine}
\alias{processRspCode}
\alias{processRspCode.RspEngine}
\alias{pushRedirect}
\alias{pushRedirect.MultiReporter}
\alias{pushRedirect.Reporter}
\alias{pushTag}
\alias{read}
\alias{read.FilterInputStream}
\alias{read.LineNumberInputStream}
\alias{readByte}
\alias{readInt}
\alias{readShort}
\alias{readText}
\alias{readText.ConnectionInputStream}
\alias{readUTF}
\alias{removeUps}
\alias{removeUps.File}
\alias{reset}
\alias{reset.FilterInputStream}
\alias{reset.LineNumberInputStream}
\alias{RspException}
\alias{setAuthor}
\alias{setAuthor.MultiReporter}
\alias{setCreator}
\alias{setDate}
\alias{setDate.MultiReporter}
\alias{setFigureNameFormat}
\alias{setFigureNameFormat.MultiReporter}
\alias{setFigurePath}
\alias{setFigurePath.MultiReporter}
\alias{setIncludePath}
\alias{setLineNumber}
\alias{setLineNumber.LineNumberInputStream}
\alias{setMask}
\alias{setOnStart}
\alias{setOnStart.LaTeXReporter}
\alias{setTitle}
\alias{setTitle.MultiReporter}
\alias{size}
\alias{skip}
\alias{skip.FilterInputStream}
\alias{split.File}
\alias{stack}
\alias{start.HtmlRotatableImage}
\alias{step}
\alias{step.default}
\alias{step.HtmlRotatableImage}
\alias{tag}
\alias{texEqToHtml}
\alias{texEqToHtml.HtmlReporter}
\alias{TextReporter}
\alias{throwException}
\alias{throwException.RspEngine}
\alias{toSafePath}
\alias{toSafePath.File}
\alias{toURL}
\alias{toURLEncodedPath}
\alias{toURLEncodedPath.File}
\alias{usepackage}
\alias{usepackage.LaTeXReporter}
\alias{write}
\alias{write.ConnectionOutputStream}
\alias{write.default}
\alias{write.FilterOutputStream}
\alias{write.MultiReporter}
\alias{writeAndEvaluateCode}
\alias{writeBold}
\alias{writeBold.HtmlReporter}
\alias{writeBold.LaTeXReporter}
\alias{writeBold.MultiReporter}
\alias{writeBold.TextReporter}
\alias{writeByte}
\alias{writeComment}
\alias{writeComment.HtmlReporter}
\alias{writeComment.LaTeXReporter}
\alias{writeComment.MultiReporter}
\alias{writeComment.TextReporter}
\alias{writeCurrentImage}
\alias{writeCurrentImage.HtmlRotatableImage}
\alias{writeDebug}
\alias{writeDebug.HtmlReporter}
\alias{writeDebug.LaTeXReporter}
\alias{writeDebug.MultiReporter}
\alias{writeDebug.TextReporter}
\alias{writeDocType}
\alias{writeEmphased}
\alias{writeEmphased.HtmlReporter}
\alias{writeEmphased.LaTeXReporter}
\alias{writeEmphased.MultiReporter}
\alias{writeEmphased.TextReporter}
\alias{writeEquation}
\alias{writeEquation.HtmlReporter}
\alias{writeEquation.LaTeXReporter}
\alias{writeEquation.MultiReporter}
\alias{writeEquation.TextReporter}
\alias{writeError}
\alias{writeError.HtmlReporter}
\alias{writeError.LaTeXReporter}
\alias{writeError.MultiReporter}
\alias{writeError.Reporter}
\alias{writeFigure}
\alias{writeFigure.HtmlReporter}
\alias{writeFigure.MultiReporter}
\alias{writeHorizontalRuler}
\alias{writeHorizontalRuler.HtmlReporter}
\alias{writeHorizontalRuler.LaTeXReporter}
\alias{writeHorizontalRuler.MultiReporter}
\alias{writeHorizontalRuler.TextReporter}
\alias{writeImage}
\alias{writeImage.HtmlReporter}
\alias{writeImage.LaTeXReporter}
\alias{writeImage.MultiReporter}
\alias{writeImage.Reporter}
\alias{writeImage.TextReporter}
\alias{writeInt}
\alias{writeItalic}
\alias{writeItalic.HtmlReporter}
\alias{writeItalic.LaTeXReporter}
\alias{writeItalic.MultiReporter}
\alias{writeItalic.TextReporter}
\alias{writeList}
\alias{writeList.HtmlReporter}
\alias{writeList.LaTeXReporter}
\alias{writeList.MultiReporter}
\alias{writeList.TextReporter}
\alias{writeNewLine}
\alias{writeNewLine.HtmlReporter}
\alias{writeNewLine.LaTeXReporter}
\alias{writeNewLine.MultiReporter}
\alias{writeNewLine.TextReporter}
\alias{writeParagraph}
\alias{writeParagraph.HtmlReporter}
\alias{writeParagraph.LaTeXReporter}
\alias{writeParagraph.MultiReporter}
\alias{writeParagraph.TextReporter}
\alias{writeSection}
\alias{writeSection.HtmlReporter}
\alias{writeSection.LaTeXReporter}
\alias{writeSection.MultiReporter}
\alias{writeSection.TextReporter}
\alias{writeShort}
\alias{writeStart}
\alias{writeStart.HtmlReporter}
\alias{writeStart.LaTeXReporter}
\alias{writeStart.TextReporter}
\alias{writeStop}
\alias{writeStop.HtmlReporter}
\alias{writeStop.LaTeXReporter}
\alias{writeStop.TextReporter}
\alias{writeSubsection}
\alias{writeSubsection.HtmlReporter}
\alias{writeSubsection.LaTeXReporter}
\alias{writeSubsection.MultiReporter}
\alias{writeSubsection.TextReporter}
\alias{writeSubsubsection}
\alias{writeSubsubsection.HtmlReporter}
\alias{writeSubsubsection.LaTeXReporter}
\alias{writeSubsubsection.MultiReporter}
\alias{writeSubsubsection.TextReporter}
\alias{writeTable}
\alias{writeTable.HtmlReporter}
\alias{writeTable.LaTeXReporter}
\alias{writeTable.MultiReporter}
\alias{writeTable.TextReporter}
\alias{writeText}
\alias{writeText.HtmlReporter}
\alias{writeText.LaTeXReporter}
\alias{writeText.MultiReporter}
\alias{writeText.TextReporter}
\alias{writeTitle}
\alias{writeTitle.HtmlReporter}
\alias{writeTitle.LaTeXReporter}
\alias{writeTitle.MultiReporter}
\alias{writeTitle.TextReporter}
\alias{writeUTF}
\alias{writeVerbatim}
\alias{writeVerbatim.HtmlReporter}
\alias{writeVerbatim.LaTeXReporter}
\alias{writeVerbatim.MultiReporter}
\alias{writeVerbatim.TextReporter}
\alias{writeWarning}
\alias{writeWarning.HtmlReporter}
\alias{writeWarning.LaTeXReporter}
\alias{writeWarning.MultiReporter}
\alias{writeWarning.Reporter}

\description{
This page contains aliases for all "non-documented" objects that
\code{R CMD check} detects in this package.

Almost all of them are \emph{generic} functions that have specific
document for the corresponding method coupled to a specific class.
Other functions are re-defined by \code{setMethodS3()} to
\emph{default} methods. Neither of these two classes are non-documented
in reality.
The rest are deprecated methods.
}

\author{Henrik Bengtsson (\url{http://www.braju.com/R/})}


\keyword{documentation}
\keyword{internal}
