%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  File.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

 \name{print.File}
\alias{print.File}
\alias{File.print}
\alias{print.File}
\alias{print,File-method}


 \title{Prints information about the File object}

 \usage{\method{print}{File}(x, ...)}

 \description{
  Prints information about the File object.
 }

 \value{
   Returns nothing.
 }



 \author{Henrik Bengtsson (\url{http://www.braju.com/R/})}

 \seealso{
   \code{\link[R.io:as.character.File]{*as.character}()}.
   For more information see \code{\link{File}}.
 }
\keyword{internal}
\keyword{methods}
