%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  HtmlPrintStream.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{printOl.HtmlPrintStream}
\alias{printOl.HtmlPrintStream}
\alias{HtmlPrintStream.printOl}
\alias{printOl.HtmlPrintStream}
\alias{printOl,HtmlPrintStream-method}


\title{Prints an object as a HTML ordered list}

\usage{\method{printOl}{HtmlPrintStream}(this, ...)}

\arguments{
  \item{...}{Arguments passed to \code{printList()}.}
}

\description{
 A wrapper function for \code{printList()} for writing ordered HTML lists.
 For more information see \code{\link[R.io:printList.HtmlPrintStream]{*printList}()}.
}

\value{
  Returns nothing.
}

\author{Henrik Bengtsson (\url{http://www.braju.com/R/})}

\seealso{
   \code{\link[R.io:printList.HtmlPrintStream]{*printList}()},
   \code{\link[R.io:printUl.HtmlPrintStream]{*printUl}()},
   \code{\link[R.io:printDl.HtmlPrintStream]{*printDl}()}, and
   \code{\link[R.io:printTable.HtmlPrintStream]{*printTable}()}.
}
\keyword{internal}
\keyword{methods}
