%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  HtmlPrintStream.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{printUl.HtmlPrintStream}
\alias{printUl.HtmlPrintStream}
\alias{HtmlPrintStream.printUl}
\alias{printUl.HtmlPrintStream}
\alias{printUl,HtmlPrintStream-method}


\title{Prints an object as a HTML unordered list}

\usage{\method{printUl}{HtmlPrintStream}(this, ...)}

\arguments{
  \item{...}{Arguments passed to \code{printList()}.}
}

\description{
 A wrapper function for \code{printList()} for writing unordered HTML lists.
 For more information see \code{\link[R.io:printList.HtmlPrintStream]{*printList}()}.
}

\value{
  Returns nothing.
}

\author{Henrik Bengtsson (\url{http://www.braju.com/R/})}

\seealso{
   \code{\link[R.io:printList.HtmlPrintStream]{*printList}()},
   \code{\link[R.io:printOl.HtmlPrintStream]{*printOl}()},
   \code{\link[R.io:printDl.HtmlPrintStream]{*printDl}()}, and
   \code{\link[R.io:printTable.HtmlPrintStream]{*printTable}()}.
}
\keyword{internal}
\keyword{methods}
