%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  999.package.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{R.io-package}
\alias{R.io-package}
\alias{R.io}
\docType{package}

\title{Package R.io}


\description{
  Contains File, connection and I/O classes.
}

\section{Installation}{
  To install this package, see \url{http://www.braju.com/R/}.
  Required packages are installed in the same way.
}

\section{Further readings}{
  Not available.
}

\section{How to cite this package}{
  Not available.
}

\author{Henrik Bengtsson (\url{http://www.braju.com/R/})}
\keyword{package}
