%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  Reporter.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{setAuthor.Reporter}
\alias{setAuthor.Reporter}
\alias{Reporter.setAuthor}
\alias{setAuthor.Reporter}
\alias{setAuthor,Reporter-method}


\title{Sets the author of the document}

\usage{\method{setAuthor}{Reporter}(this, author=getOption("author"), ..., collapse="", sep="")}

\arguments{
  \item{author}{the author of the document.}
}

\description{
  Sets the author of the document.
}

\value{
  Returns nothing.
}

\author{Henrik Bengtsson (\url{http://www.braju.com/R/})}

\seealso{
  For more information see \code{\link{Reporter}}.
}
\keyword{internal}
\keyword{methods}
