%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  Reporter.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{setDate.Reporter}
\alias{setDate.Reporter}
\alias{Reporter.setDate}
\alias{setDate.Reporter}
\alias{setDate,Reporter-method}


\title{Sets the date of the document}

\usage{\method{setDate}{Reporter}(this, date=NULL, ...)}

\arguments{
  \item{date}{the date of the document. If \code{\link[base]{NULL}}, the current date is used.}
}

\description{
  Sets the date of the document.
}

\value{
  Returns nothing.
}

\author{Henrik Bengtsson (\url{http://www.braju.com/R/})}

\seealso{
  For more information see \code{\link{Reporter}}.
}
\keyword{internal}
\keyword{methods}
