%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  Reporter.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{setTitle.Reporter}
\alias{setTitle.Reporter}
\alias{Reporter.setTitle}
\alias{setTitle.Reporter}
\alias{setTitle,Reporter-method}


\title{Sets the title of the document}

\usage{\method{setTitle}{Reporter}(this, ..., collapse="", sep="")}

\arguments{
  \item{title}{the title of the document.}
}

\description{
  Sets the title of the document.
}

\value{
  Returns nothing.
}

\author{Henrik Bengtsson (\url{http://www.braju.com/R/})}

\seealso{
  For more information see \code{\link{Reporter}}.
}
\keyword{internal}
\keyword{methods}
