%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  File.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

 \name{show.File}
\alias{show.File}
\alias{File.show}
\alias{show.File}
\alias{show,File-method}


 \title{Shows the current file}

 \usage{\method{show}{File}(this, format=NULL, ...)}

 \description{
  Shows the current file by calling the correct viewer/player based on the file name
  extension.
 }

 \value{
  Returns nothing.
 }



 \author{Henrik Bengtsson (\url{http://www.braju.com/R/})}

 \seealso{
   For more information see \code{\link{File}}.
 }
\keyword{internal}
\keyword{methods}
