%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  File.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

 \name{toURL.File}
\alias{toURL.File}
\alias{File.toURL}
\alias{toURL.File}
\alias{toURL,File-method}


 \title{Converts this abstract pathname into a URL}

 \usage{\method{toURL}{File}(this, safe=TRUE, ...)}

 \description{
  Converts this abstract pathname into a URL starting with \code{file://}.
 }

 \value{
  Returns a \code{\link[base]{character}} string.
 }



 \author{Henrik Bengtsson (\url{http://www.braju.com/R/})}

 \seealso{
   For more information see \code{\link{File}}.
 }
\keyword{internal}
\keyword{methods}
