%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  BufferedOutputStream.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{write.BufferedOutputStream}
\alias{write.BufferedOutputStream}
\alias{BufferedOutputStream.write}
\alias{write.BufferedOutputStream}
\alias{write,BufferedOutputStream-method}


\title{Writes one or more bytes to the output stream}

\usage{\method{write}{BufferedOutputStream}(this, b, off=1, len=length(b), ...)}

\description{
 Writes one or more bytes to the output stream.
}

\value{
  Returns nothing.
}

\author{Henrik Bengtsson (\url{http://www.braju.com/R/})}

\seealso{
  For more information see \code{\link{BufferedOutputStream}}..
}
\keyword{internal}
\keyword{methods}
