%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  FileOutputStream.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{write.FileOutputStream}
\alias{write.FileOutputStream}
\alias{FileOutputStream.write}
\alias{write.FileOutputStream}
\alias{write,FileOutputStream-method}


\title{Writes one or more bytes to the file output stream}

\usage{\method{write}{FileOutputStream}(this, b, off=1, len=length(b), ...)}

\description{
 Writes one or more bytes to the file output stream.
}

\value{
  Returns nothing.
}

\author{Henrik Bengtsson (\url{http://www.braju.com/R/})}

\seealso{
 For more information see \code{\link{FileOutputStream}}.
}
\keyword{internal}
\keyword{methods}
