%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  MultiOutputStream.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{write.MultiOutputStream}
\alias{write.MultiOutputStream}
\alias{MultiOutputStream.write}
\alias{write.MultiOutputStream}
\alias{write,MultiOutputStream-method}


\title{Writes one or more bytes to all output streams}

\usage{\method{write}{MultiOutputStream}(this, b, off=1, len=length(b), ...)}

\description{
 Writes one or more bytes to all output streams.
}

\value{
  Returns nothing.
}

\author{Henrik Bengtsson (\url{http://www.braju.com/R/})}

\seealso{
  For more information see \code{\link{MultiOutputStream}}.
}
\keyword{internal}
\keyword{methods}
