%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  OutputStream.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{write.OutputStream}
\alias{write.OutputStream}
\alias{OutputStream.write}
\alias{write.OutputStream}
\alias{write,OutputStream-method}


\title{Writes one or more bytes to the output stream}

\usage{\method{write}{OutputStream}(...)}

\description{
 Writes one or more bytes to the output stream.
}

\value{
  Returns nothing.
}

\seealso{
  For more information see \code{\link{OutputStream}}.
}

\author{Henrik Bengtsson (\url{http://www.braju.com/R/})}
\keyword{internal}
\keyword{methods}
