%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  Reporter.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

 \name{writeAndEvaluateCode.Reporter}
\alias{writeAndEvaluateCode.Reporter}
\alias{Reporter.writeAndEvaluateCode}
\alias{writeAndEvaluateCode.Reporter}
\alias{writeAndEvaluateCode,Reporter-method}


 \title{Writes and then evaluates the code}

 \usage{\method{writeAndEvaluateCode}{Reporter}(this, ..., envir=parent.frame())}

 \arguments{
   \item{...}{Expressions to be written and evaluated.}
   \item{envir}{Environment where the expressions should be evaluated.}
 }

 \description{
   Writes and then evaluates the code.
 }

 \author{Henrik Bengtsson (\url{http://www.braju.com/R/})}

 \seealso{
   For more information see \code{\link{Reporter}}.
 }
\keyword{internal}
\keyword{methods}
