%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  Reporter.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{writeBold.Reporter}
\alias{writeBold.Reporter}
\alias{Reporter.writeBold}
\alias{writeBold.Reporter}
\alias{writeBold,Reporter-method}


\title{Writes a sentence in bold face font}

\usage{\method{writeBold}{Reporter}(this, ...)}

\description{
  Writes a sentence in bold face font.
}

\author{Henrik Bengtsson (\url{http://www.braju.com/R/})}

\seealso{
  For more information see \code{\link{Reporter}}.
}
\keyword{internal}
\keyword{methods}
