%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  Reporter.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{writeComment.Reporter}
\alias{writeComment.Reporter}
\alias{Reporter.writeComment}
\alias{writeComment.Reporter}
\alias{writeComment,Reporter-method}


\title{Writes a comment not visible in the final document}

\usage{\method{writeComment}{Reporter}(this, ...)}

\description{
  Writes a comment not visible in the final document.
}

\author{Henrik Bengtsson (\url{http://www.braju.com/R/})}

\seealso{
  For more information see \code{\link{Reporter}}.
}
\keyword{internal}
\keyword{methods}
