%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  Reporter.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{writeDebug.Reporter}
\alias{writeDebug.Reporter}
\alias{Reporter.writeDebug}
\alias{writeDebug.Reporter}
\alias{writeDebug,Reporter-method}


\title{Writes debug information}

\usage{\method{writeDebug}{Reporter}(this, ...)}

\description{
  Writes debug information.
}

\author{Henrik Bengtsson (\url{http://www.braju.com/R/})}

\seealso{
  For more information see \code{\link{Reporter}}.
}
\keyword{internal}
\keyword{methods}
