%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  HtmlPrintStream.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{writeDocType.HtmlPrintStream}
\alias{writeDocType.HtmlPrintStream}
\alias{HtmlPrintStream.writeDocType}
\alias{writeDocType.HtmlPrintStream}
\alias{writeDocType,HtmlPrintStream-method}


\title{Writes the document type string}

\usage{\method{writeDocType}{HtmlPrintStream}(this, type="HTML 4.0 Transitional", ...)}

\arguments{
  \item{type}{The document type string to write. Default value is
    \code{"-//W3C//DTD HTML 4.01 Transitional//EN"}.}
}

\description{
 Writes the document type string to the connected output stream.
}

\value{
  Returns nothing.
}

\author{Henrik Bengtsson (\url{http://www.braju.com/R/})}
\keyword{internal}
\keyword{methods}
