%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  Reporter.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{writeEmphased.Reporter}
\alias{writeEmphased.Reporter}
\alias{Reporter.writeEmphased}
\alias{writeEmphased.Reporter}
\alias{writeEmphased,Reporter-method}


\title{Writes a sentence in an emphasized font}

\usage{\method{writeEmphased}{Reporter}(this, ...)}

\description{
  Writes a sentence in an emphasized font.
}

\author{Henrik Bengtsson (\url{http://www.braju.com/R/})}

\seealso{
  For more information see \code{\link{Reporter}}.
}
\keyword{internal}
\keyword{methods}
