%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  Reporter.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{writeEquation.Reporter}
\alias{writeEquation.Reporter}
\alias{Reporter.writeEquation}
\alias{writeEquation.Reporter}
\alias{writeEquation,Reporter-method}


\title{Writes an equation}

\usage{\method{writeEquation}{Reporter}(this, ...)}

\description{
  Writes an equation.
}

\author{Henrik Bengtsson (\url{http://www.braju.com/R/})}

\seealso{
  For more information see \code{\link{Reporter}}.
}
\keyword{internal}
\keyword{methods}
