%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  Reporter.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{writeFigure.Reporter}
\alias{writeFigure.Reporter}
\alias{Reporter.writeFigure}
\alias{writeFigure.Reporter}
\alias{writeFigure,Reporter-method}


\title{Writes a figure by saving the current plot}

\usage{\method{writeFigure}{Reporter}(this, label=NULL, caption="", ...)}

\description{
  Writes a figure by saving the current plot.
}

\author{Henrik Bengtsson (\url{http://www.braju.com/R/})}

\seealso{
  For more information see \code{\link{Reporter}}.
}
\keyword{internal}
\keyword{methods}
