%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  Reporter.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{writeHorizontalRuler.Reporter}
\alias{writeHorizontalRuler.Reporter}
\alias{Reporter.writeHorizontalRuler}
\alias{writeHorizontalRuler.Reporter}
\alias{writeHorizontalRuler,Reporter-method}


\title{Adds a horizontal ruler to the document}

\usage{\method{writeHorizontalRuler}{Reporter}(...)}

\description{
  Adds a horizontal ruler to the document.
}

\author{Henrik Bengtsson (\url{http://www.braju.com/R/})}

\seealso{
  For more information see \code{\link{Reporter}}.
}
\keyword{internal}
\keyword{methods}
