%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  Reporter.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{writeItalic.Reporter}
\alias{writeItalic.Reporter}
\alias{Reporter.writeItalic}
\alias{writeItalic.Reporter}
\alias{writeItalic,Reporter-method}


\title{Writes a sentence in an italic font}

\usage{\method{writeItalic}{Reporter}(this, ...)}

\description{
  Writes a sentence in an italic font.
}

\author{Henrik Bengtsson (\url{http://www.braju.com/R/})}

\seealso{
  For more information see \code{\link{Reporter}}.
}
\keyword{internal}
\keyword{methods}
