%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  javaFunctions.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{writeJavaUTF}
\alias{writeJavaUTF.default}
\alias{writeJavaUTF}

\title{Deprecated. Writes a string to a connection in Java format (UTF-8)}

\description{
  Writes a string to a connection in Java format (UTF-8)
  so it will be readable by Java.
}

\usage{\method{writeJavaUTF}{default}(con, str, ...)}

\arguments{
  \item{con}{Binary connection to be written to.}
  \item{str}{String to be written.}
}

\author{Henrik Bengtsson (\url{http://www.braju.com/R/})}

\seealso{
  \code{\link[base]{writeBin}}().
  \code{\link{writeJavaShort}}(), \code{\link{writeJavaInt}}(),
  \code{\link{writeJavaUTF}}().
}

\keyword{internal}

