%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  Reporter.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{writeNewLine.Reporter}
\alias{writeNewLine.Reporter}
\alias{Reporter.writeNewLine}
\alias{writeNewLine.Reporter}
\alias{writeNewLine,Reporter-method}


\title{Adds a new line/carrage return to the document}

\usage{\method{writeNewLine}{Reporter}(...)}

\description{
  Adds a new line/carrage return to the document.
}

\author{Henrik Bengtsson (\url{http://www.braju.com/R/})}

\seealso{
  For more information see \code{\link{Reporter}}.
}
\keyword{internal}
\keyword{methods}
