%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  Reporter.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{writeParagraph.Reporter}
\alias{writeParagraph.Reporter}
\alias{Reporter.writeParagraph}
\alias{writeParagraph.Reporter}
\alias{writeParagraph,Reporter-method}


\title{Writes a paragraph}

\usage{\method{writeParagraph}{Reporter}(this, ...)}

\description{
  Writes a paragraph.
}

\author{Henrik Bengtsson (\url{http://www.braju.com/R/})}

\seealso{
  For more information see \code{\link{Reporter}}.
}
\keyword{internal}
\keyword{methods}
