%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  Reporter.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{writeSubsection.Reporter}
\alias{writeSubsection.Reporter}
\alias{Reporter.writeSubsection}
\alias{writeSubsection.Reporter}
\alias{writeSubsection,Reporter-method}


\title{Begins a subsection}

\usage{\method{writeSubsection}{Reporter}(...)}

\description{
  Begins a subsection.
}

\author{Henrik Bengtsson (\url{http://www.braju.com/R/})}

\seealso{
  For more information see \code{\link{Reporter}}.
}
\keyword{internal}
\keyword{methods}
