%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  Reporter.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{writeSubsubsection.Reporter}
\alias{writeSubsubsection.Reporter}
\alias{Reporter.writeSubsubsection}
\alias{writeSubsubsection.Reporter}
\alias{writeSubsubsection,Reporter-method}


\title{Begins a subsubsection}

\usage{\method{writeSubsubsection}{Reporter}(...)}

\description{
  Begins a subsubsection.
}

\author{Henrik Bengtsson (\url{http://www.braju.com/R/})}

\seealso{
  For more information see \code{\link{Reporter}}.
}
\keyword{internal}
\keyword{methods}
