%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  Reporter.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{writeVerbatim.Reporter}
\alias{writeVerbatim.Reporter}
\alias{Reporter.writeVerbatim}
\alias{writeVerbatim.Reporter}
\alias{writeVerbatim,Reporter-method}


\title{Writes an object in verbatim}

\usage{\method{writeVerbatim}{Reporter}(this, ...)}

\description{
  Writes an object in verbatim.
}

\author{Henrik Bengtsson (\url{http://www.braju.com/R/})}

\seealso{
  For more information see \code{\link{Reporter}}.
}
\keyword{internal}
\keyword{methods}
